requirements.txt
# PolyBuk Framework Dependencies
# Install: pip install -r requirements.txt

# --- Polymarket ---
py-clob-client>=0.18.0       # Official Polymarket CLOB API SDK (orders, orderbook)
web3>=6.0.0                  # Ethereum signing (required by py-clob-client for auth)

# --- Database ---
supabase>=2.0.0              # Official Supabase Python SDK

# --- Telegram Alerts ---
python-telegram-bot>=21.0    # Telegram bot (alerts, /kill, /status commands)

# --- HTTP & Environment ---
httpx>=0.27.0                # Async HTTP client for Gamma API calls
python-dotenv>=1.0.0         # Load .env file into environment variables
